functionnal snapshot

+++ src/main/java/com/ordwen/simpleblockfreeze/storage/sql/ManageLocation.java
@@ -1,6 +1,6 @@
 package com.ordwen.simpleblockfreeze.storage.sql;
 
-import com.ordwen.simpleblockfreeze.Messages;
+import com.ordwen.simpleblockfreeze.enums.Messages;
 import com.ordwen.simpleblockfreeze.tools.PluginLogger;
 import org.bukkit.entity.Player;
 
@@ -31,19 +31,20 @@ public class ManageLocation {
      * @param z     z coordinate.
      */
     public void saveLocation(Player player, String world, double x, double y, double z) {
-        final Connection connection = sqlManager.getConnection();
         if (sqlManager.searchLocation(world, x, y, z)) {
             final String msg = Messages.ALREADY_FROZEN.toString();
             if (msg != null) player.sendMessage(msg);
             return;
         }
 
+        final Connection connection = sqlManager.getConnection();
         try {
             final String SAVE_QUERY = "INSERT INTO SBF_LOCATIONS(WORLD_NAME, X, Y, Z) VALUES (?, ?, ?, ?)";
             getStatement(world, x, y, z, connection, SAVE_QUERY);
 
             final String msg = Messages.FREEZE_SUCCESS.toString();
             if (msg != null) player.sendMessage(msg);
+            connection.close();
         } catch (SQLException e) {
             PluginLogger.error("An error occurred while saving a block location in the database.");
             PluginLogger.error(e.getMessage());
@@ -62,14 +63,20 @@ public class ManageLocation {
      * @param z     z coordinate.
      */
     public void deleteLocation(Player player, String world, double x, double y, double z) {
-        final Connection connection = sqlManager.getConnection();
+        if (!sqlManager.searchLocation(world, x, y, z)) {
+            final String msg = Messages.FREEZE_NOT_FOUND.toString();
+            if (msg != null) player.sendMessage(msg);
+            return;
+        }
 
+        final Connection connection = sqlManager.getConnection();
         try {
             final String DELETE_QUERY = "DELETE FROM SBF_LOCATIONS WHERE WORLD_NAME = ? AND X = ? AND Y = ? AND Z = ?";
             getStatement(world, x, y, z, connection, DELETE_QUERY);
 
             final String msg = Messages.UNFREEZE_SUCCESS.toString();
             if (msg != null) player.sendMessage(msg);
+            connection.close();
         } catch (SQLException e) {
             PluginLogger.error("An error occurred while deleting a block location in the database.");
             PluginLogger.error(e.getMessage());
@@ -100,6 +107,7 @@ public class ManageLocation {
         playerStatement.setDouble(4, z);
 
         playerStatement.executeUpdate();
+        playerStatement.close();
         connection.close();
     }
 }
